requirements.txt
gymnasium>=0.29
minigrid>=2.3
numpy>=1.24
torch>=2.1
transformers>=4.45
accelerate>=0.26
sentencepiece>=0.1
pandas>=2.0
matplotlib>=3.7
tqdm>=4.66
pyyaml>=6.0
